Sport-Data-Hub/pyproject.toml
[tool.poetry]
name = "sport-data-hub"
version = "0.1.0"
description = ""
authors = ["junchen88 <jimmy181820@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
